refactor(tui): split agenda_model.go below 1000 lines
Move update and view into sibling files in package tui.
No behavior change.

Related: #652
